Update API
health was changed to a float
networkclass was obsolete

+++ code/Util/PlayerUtil.cs
@@ -17,7 +17,7 @@ namespace CombineControl.Util
 		{
 			if ( ConsoleSystem.Caller.Pawn is CCPlayer player )
 			{
-				int oldHealth = player.Health;
+				float oldHealth = player.Health;
 				player.TakeDamage( DamageInfo.FromBullet( player.Position, Vector3.Zero, damage ) );
 
 				Log.Info( $"Old health: {oldHealth}, new health: {player.Health}" );
